Revert "search, predictionについてNotifierProviderに変更"
