+++ lib/pages/search/search_header.dart
@@ -1,9 +1,8 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../../Controller/provider/search/placeapi_provider.dart';
-import '../../Controller/provider/search/prediction_provider.dart';
+import 'package:kasairu/controller/provider/search/placeapi_provider.dart';
+import 'package:kasairu/controller/provider/search/prediction_provider.dart';
 
 class SearchScreenHeader extends ConsumerStatefulWidget
     implements PreferredSizeWidget {
@@ -23,10 +22,10 @@ class SearchScreenHeaderState extends ConsumerState {
   Timer? debounce;
 
   void _autocomplete(text) async {
-    var api = ref.watch(placeAPIProvider);
-    var result = await api.getAPI().autocomplete.get(text);
+    final api = ref.watch(placeAPIProvider);
+    final result = await api.getAutoComplete(text);
     if (result != null && result.predictions != null && mounted) {
-      ref.read(predictionProvider.notifier).setPrediction(result.predictions!);
+      ref.read(predictionProvider.notifier).state = result.predictions!;
     }
   }
