+++ lib/pages/search/search_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:kasairu/models/place/place.dart';
 import 'package:kasairu/pages/search/search_header.dart';
 import 'package:kasairu/pages/weather/detail/search_result_screen.dart';
+import 'package:kasairu/controller/provider/search/placeapi_provider.dart';
 import 'package:kasairu/controller/provider/search/prediction_provider.dart';
 import 'package:google_place/google_place.dart';
 import 'package:kasairu/controller/provider/search/search_provider.dart';
-
-import '../../Controller/provider/search/placeapi_provider.dart';
 
 class SearchScreen extends ConsumerStatefulWidget {
   const SearchScreen({super.key});
@@ -30,13 +30,16 @@ class SearchScreenState extends ConsumerState {
     focus.dispose();
   }
 
-  void _forecastSearch(AutocompletePrediction pred) async {
+  Future<void> _forecastSearch(AutocompletePrediction pred) async {
     final placeId = pred.placeId!;
     final api = ref.watch(placeAPIProvider);
-
     final details = await api.getSearchPlace(placeId);
-    if (details != null && details.result != null) {
-      ref.read(searchProvider.notifier).setSearchPlace(details);
+    if (details != null && details.result != null && mounted) {
+      var lat = details.result!.geometry!.location!.lat;
+      var lng = details.result!.geometry!.location!.lng;
+      Place searchPlace =
+          Place(name: details.result!.name!, lat: lat!, lng: lng!);
+      ref.read(searchProvider.notifier).state = searchPlace;
     }
   }
 
@@ -77,8 +80,8 @@ class SearchScreenState extends ConsumerState {
                           dense: true,
                           title: Text(pred.description.toString()),
                           minLeadingWidth: 4.0,
-                          onTap: () {
-                            _forecastSearch(pred);
+                          onTap: () async {
+                            await _forecastSearch(pred);
                             if (!mounted) return;
                             Navigator.pushReplacement(
                               context,
