+++ lib/repository/api/openweather/weather_api.dart
@@ -7,16 +7,10 @@ import 'package:kasairu/models/onecall/onecall_response.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 
-const host = "https://api.openweathermap.org";
-const pathOneCall = "/data/3.0/onecall";
-const exclude = "current,minutely,alerts";
-
 // APIを呼び出し、データを取得する
 Future<OneCallResponse> fetchOneCall(Place loc) async {
-  const key = String.fromEnvironment("OPENWEATHER_KEY");
-  final String uri = "$host$pathOneCall?lat=${loc.lat}&lon=${loc.lng}&exclude=$exclude&appid=$key";
-  debugPrint(uri);
-
+  APIQuery apiQuery = APIQuery();
+  final uri = await apiQuery.openOneCallQuery(loc);
   try {
     final response = await http.get(Uri.parse(uri));
     if (response.statusCode == 200) {
@@ -27,5 +21,21 @@ Future<OneCallResponse> fetchOneCall(Place loc) async {
     }
   } catch (error) {
     throw Exception(error.toString());
+  }
+}
+
+class APIQuery {
+  static const host = "https://api.openweathermap.org";
+  static const pathOneCall = "/data/3.0/onecall";
+  static const exclude = "current,minutely,alerts";
+
+  Future<String> openOneCallQuery(Place loc) async {
+    var key = const String.fromEnvironment("OPENWEATHER_KEY");
+    String uri =
+        "$host$pathOneCall?lat=${loc.lat}&lon=${loc.lng}&exclude=$exclude&appid=$key";
+
+    debugPrint(uri);
+
+    return uri;
   }
 }
