+++ lib/repository/settings/settings_repository.dart
@@ -26,4 +26,6 @@ class SettingsRepository {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     await prefs.setInt(_keyHourThreshold, value);
   }
+
+  getFavorite() {}
 }
