+++ pubspec.yaml
@@ -44,8 +44,6 @@ dependencies:
   flutter_dotenv: ^5.0.2
   settings_ui: ^2.0.2
   path: ^1.8.3
-  riverpod_lint: ^2.3.9
-  riverpod_generator: ^2.3.11
   flutter:
     sdk: flutter
